+++ Src/Commands/GetDataCommand.cs
@@ -3,54 +3,100 @@
 // http://www.apache.org/licenses/LICENSE-2.0
 
 using LiteDB;
+using System;
 using System.Management.Automation;
 
 namespace Ldbc.Commands
 {
-	[Cmdlet(VerbsCommon.Get, "LiteData")]
+	[Cmdlet(VerbsCommon.Get, "LiteData", DefaultParameterSetName = nsData)]
 	[OutputType(typeof(Dictionary))]
 	public sealed class GetDataCommand : Abstract
 	{
+		const string nsData = "Data";
+		const string nsCount = "Count";
+
 		[Parameter(Position = 0, Mandatory = true)]
 		public ILiteCollection<BsonDocument> Collection { get; set; }
 
 		[Parameter(Position = 1)]
-		public object[] Filter { set { _Filter = Expression.Create(value); } }
+		public object Filter { set { _Filter = Expression.Create(value); } }
 		Expression _Filter;
 
-		[Parameter]
+		[Parameter(Mandatory = true, ParameterSetName = nsCount)]
 		public SwitchParameter Count { get; set; }
 
-		void DoCount()
+		[Parameter(ParameterSetName = nsData)]
+		public int First { get; set; }
+
+		[Parameter(ParameterSetName = nsData)]
+		public int Last { get; set; }
+
+		[Parameter(ParameterSetName = nsData)]
+		public int Skip { get; set; }
+
+		[Parameter(ParameterSetName = nsData)]
+		public object As { get; set; }
+
+		int GetCount()
 		{
-			if (_Filter == null)
-			{
-				WriteObject(Collection.Count());
-			}
-			else
-			{
-				WriteObject(Collection.Count(_Filter.BsonExpression));
-			}
+			return _Filter == null ? Collection.Count() : Collection.Count(_Filter.BsonExpression);
+		}
+
+		bool DoLast()
+		{
+			if (Last <= 0)
+				return false;
+
+			Skip = GetCount() - Skip - Last;
+			First = Last;
+			if (Skip >= 0)
+				return false;
+
+			First += Skip;
+			if (First <= 0)
+				return true;
+
+			Skip = 0;
+			return false;
 		}
 
 		protected override void BeginProcessing()
 		{
+			if (First > 0 && Last > 0)
+				throw new PSArgumentException("Parameters First and Last cannot be specified together.");
+
 			if (Count)
 			{
-				DoCount();
+				WriteObject(GetCount());
 				return;
 			}
 
-			if (_Filter == null)
+			if (DoLast())
+				return;
+
+			var query = Collection.Query();
+			if (_Filter != null)
+				query = query.Where(_Filter.BsonExpression);
+
+			var result = query.Skip(Skip);
+			if (First > 0)
+				result = result.Limit(First);
+
+			if (As == null)
 			{
-				foreach (var doc in Collection.FindAll())
+				foreach (var doc in result.ToEnumerable())
 					WriteObject(Actor.ToObject(doc));
+				return;
 			}
-			else
+
+			if (As is string s && s.Equals("PS", StringComparison.OrdinalIgnoreCase))
 			{
-				foreach (var doc in Collection.Find(_Filter.BsonExpression))
-					WriteObject(Actor.ToObject(doc));
+				foreach (var doc in result.ToEnumerable())
+					WriteObject(PSObjectSeializer.ReadCustomObject(doc));
+				return;
 			}
+
+			throw new PSNotSupportedException("Parameter As must be just 'PS', for now.");
 		}
 	}
 }