+++ Tests/Add.test.ps1
@@ -25,24 +25,24 @@ task Result {
 	}
 }
 
-task Batch {
+task Bulk {
 	Use-LiteDatabase :memory: {
 		$test = Get-LiteCollection Test
 
 		try {
-			# add batch with duplicated keys
-			@{_id=1}, @{_id=1} | Add-LiteData $test -Batch
+			# add bulk with duplicated keys
+			@{_id=1}, @{_id=1} | Add-LiteData $test -Bulk
 			throw
 		}
 		catch {
 			equals "$_" "Cannot insert duplicate key in unique index '_id'. The duplicate value is '1'."
 		}
 
-		# batch is undone
+		# bulk is undone
 		equals $test.Count() 0
 
-		# add normal batch
-		$r = @{_id=1}, @{_id=2} | Add-LiteData $test -Batch -Result
+		# add normal bulk
+		$r = @{_id=1}, @{_id=2} | Add-LiteData $test -Bulk -Result
 		equals $r 2
 		equals $test.Count() 2
 	}