+++ Tests/Get.test.ps1
@@ -1,11 +1,20 @@
 
-Import-Module Ldbc
+. ./Zoo.ps1
+
+task AsPSCustomObject {
+	Use-LiteDatabase :memory: {
+		$test = Get-LiteCollection Test Int32
+		Add-LiteData $test (New-BsonBag)
+		$r = Get-LiteData $test -As ps
+		equals $r.GetType().Name PSCustomObject
+	}
+}
 
 task CountAndExists {
 	Use-LiteDatabase :memory: {
 		$test = Get-LiteCollection Test Int32
 
-		# add data for tests
+		# add data
 		@{name = 'John'; age = 42}, @{name = 'Mary'; age = 33} | Add-LiteData $test
 
 		# test Test-LiteData
@@ -20,5 +29,44 @@ task CountAndExists {
 		equals (Get-LiteData -Count $test '$.age > 20') 2
 		equals (Get-LiteData -Count $test '$.age > 40') 1
 		equals (Get-LiteData -Count $test '$.age > 50') 0
+	}
+}
+
+task FirstLastSkip {
+	Use-LiteDatabase :memory: {
+		$test = Get-LiteCollection Test
+
+		# add data
+		1..9 | .{process{ @{_id = $_} }} | Add-LiteData $test
+
+		# get all
+
+		$r = Get-LiteData $test -First 2
+		equals "$r" '{"_id":1} {"_id":2}'
+
+		$r = Get-LiteData $test -Last 2
+		equals "$r" '{"_id":8} {"_id":9}'
+
+		$r = Get-LiteData $test -Skip 2 -First 2
+		equals "$r" '{"_id":3} {"_id":4}'
+
+		$r = Get-LiteData $test -Skip 2 -Last 2
+		equals "$r" '{"_id":6} {"_id":7}'
+
+		# get where
+
+		$filter = '$._id >= 3 AND $._id <= 7'
+
+		$r = Get-LiteData $test $filter -First 2
+		equals "$r" '{"_id":3} {"_id":4}'
+
+		$r = Get-LiteData $test $filter -Last 2
+		equals "$r" '{"_id":6} {"_id":7}'
+
+		$r = Get-LiteData $test $filter -Skip 2 -First 2
+		equals "$r" '{"_id":5} {"_id":6}'
+
+		$r = Get-LiteData $test $filter -Skip 2 -Last 2
+		equals "$r" '{"_id":4} {"_id":5}'
 	}
 }