+++ Tests/Set.test.ps1
@@ -26,31 +26,28 @@ task Basic {
 	}
 }
 
-#TODO fixed in LiteDB master, wait, convert to batch test
-task NRE {
+task Bulk {
 	Use-LiteDatabase :memory: {
 		$test = Get-LiteCollection Test
-		try {
-			@{_id=1; n=1} | Set-LiteData $test -Batch
-			throw
-		}
-		catch {
-			$_
-		}
+
+		# add bulk (use duplicated key)
+		$r = @{_id=1}, @{_id=1; p=1}, @{_id=2; p=2} | Set-LiteData $test -Add -Bulk -Result
+		equals $r 2
+
+		# set bulk (use missing key)
+		$r = @{_id=1; p=2}, @{_id=2; p=3}, @{_id=3; p=4} | Set-LiteData $test -Bulk -Result
+		equals $r 2
+
+		$r = Get-LiteData $test
+		equals "$r" '{"_id":1,"p":2} {"_id":2,"p":3}'
 	}
 }
 
-#?? https://github.com/mbdavid/LiteDB/issues/1451
+# https://github.com/mbdavid/LiteDB/issues/1451
 task UpdateInMissingCollection {
 	Use-LiteDatabase :memory: {
 		$test = Get-LiteCollection Test
-		try {
-			$r = Set-LiteData $test @{_id = 1; v = 1} -Result
-			equals $r 0
-			throw
-		}
-		catch {
-			equals $_.FullyQualifiedErrorId 'System.NullReferenceException,Ldbc.Commands.SetDataCommand'
-		}
+		$r = Set-LiteData $test @{_id = 1; v = 1} -Result
+		equals $r 0
 	}
 }